Change int to decimal for Amount

+++ TransactionStore/TransactionStore.DB/Stored Procedures/Transaction_AddDepositOrWithdraw.sql
@@ -1,6 +1,6 @@
 ﻿CREATE PROCEDURE [dbo].[Transaction_AddDepositOrWithdraw]
 	@leadId int,
-	@amount int,
+	@amount decimal,
 	@type int
 as
 begin
